Expose GUI driver facade alias
