+++ .memory/memory/decisions/gui-driver-facade-alias.json
@@ -1,0 +1,53 @@
+{
+  "body_path": "memory/decisions/gui-driver-facade-alias.md",
+  "content_hash": "sha256:9c68aeee368ceb374808c4cd1ce08ba7d1d8d0137aab4b24a67a117af5cd3d85",
+  "created_at": "2026-06-21T12:09:46+03:00",
+  "evidence": [
+    {
+      "id": "shaft-engine/src/main/java/com/shaft/driver/SHAFT.java",
+      "kind": "file"
+    },
+    {
+      "id": "shaft-engine/src/main/java/com/shaft/gui/driver/DriverContract.java",
+      "kind": "file"
+    },
+    {
+      "id": "shaft-engine/src/test/java/testPackage/unitTests/GUIInterfaceCompatibilityCoverageUnitTest.java",
+      "kind": "file"
+    }
+  ],
+  "facets": {
+    "applies_to": [
+      "shaft-engine/src/main/java/com/shaft/driver/SHAFT.java",
+      "shaft-engine/src/main/java/com/shaft/gui/driver/DriverContract.java",
+      "shaft-engine/src/test/java/testPackage/unitTests/GUIInterfaceCompatibilityCoverageUnitTest.java"
+    ],
+    "category": "decision-rationale",
+    "load_modes": [
+      "coding",
+      "review",
+      "architecture"
+    ]
+  },
+  "id": "decision.gui-driver-facade-alias",
+  "scope": {
+    "branch": null,
+    "kind": "project",
+    "project": "project.shaft-engine",
+    "task": null
+  },
+  "source": {
+    "kind": "agent",
+    "task": "Expose generic SHAFT GUI driver facade for Selenium and Playwright"
+  },
+  "status": "active",
+  "tags": [
+    "gui",
+    "playwright",
+    "webdriver",
+    "public-api"
+  ],
+  "title": "Use SHAFT.GUI.Driver as the public GUI driver facade",
+  "type": "decision",
+  "updated_at": "2026-06-21T12:09:46+03:00"
+}
