+++ .memory/memory/decisions/gui-driver-facade-alias.md
@@ -1,0 +1,1 @@
+Generic GUI sessions should be declared as `SHAFT.GUI.Driver`, with both `SHAFT.GUI.WebDriver` and `SHAFT.GUI.Playwright` implementing that nested facade. Keep the package-level contract named `com.shaft.gui.driver.DriverContract` so code-quality checks do not flag a type with the same simple name as its supertype, and avoid requiring users to import an internal GUI driver contract.
