+++ graphify-out/.graphify_labels.json
@@ -765,5 +765,32 @@
   "763": "Community 763",
   "764": "Community 764",
   "765": "Community 765",
-  "766": "Community 766"
+  "766": "Community 766",
+  "767": "Community 767",
+  "768": "Community 768",
+  "769": "Community 769",
+  "770": "Community 770",
+  "771": "Community 771",
+  "772": "Community 772",
+  "773": "Community 773",
+  "774": "Community 774",
+  "775": "Community 775",
+  "776": "Community 776",
+  "777": "Community 777",
+  "778": "Community 778",
+  "779": "Community 779",
+  "780": "Community 780",
+  "781": "Community 781",
+  "782": "Community 782",
+  "783": "Community 783",
+  "784": "Community 784",
+  "785": "Community 785",
+  "786": "Community 786",
+  "787": "Community 787",
+  "788": "Community 788",
+  "789": "Community 789",
+  "790": "Community 790",
+  "791": "Community 791",
+  "792": "Community 792",
+  "793": "Community 793"
 }
